pyproject.toml
[project]
name = "hello-world"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.3.0",
    "diffusers>=0.32.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scikit-image>=0.25.0",
    "torch==2.5.0",
    "torchvision==0.20.0",
    "transformers>=4.48.1",
]
[tool.uv.sources]
torch = [
      { index = "pytorch-cu124", marker = "platform_system == 'Linux'" },
]
torchvision = [
{ index = "pytorch-cu124" , marker = "platform_system == 'Linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
